+++ api-module-library/ironclad/api.js
@@ -19,6 +19,10 @@ class Api extends ApiKeyRequester {
             workflowSchemas: '/public/api/v1/workflow-schemas',
             workflowSchemaByID: (schemaId) =>
                 `/public/api/v1/workflow-schemas/${schemaId}`,
+            workflowMetadata: (workflowId) =>
+                `/public/api/v1/workflows/${workflowId}/attributes`,
+            workflowComment: (workflowId) =>
+                `/public/api/v1/workflows/${workflowId}/comment`,
             records: '/public/api/v1/records',
             recordByID: (recordId) => `/public/api/v1/records/${recordId}`,
             recordSchemas: '/public/api/v1/records/metadata',
@@ -76,6 +80,7 @@ class Api extends ApiKeyRequester {
         const response = await this._patch(options);
         return response;
     }
+
     async deleteWebhook(webhookId) {
         const options = {
             url: this.baseUrl + this.URLs.webhookByID(webhookId),
@@ -84,9 +89,10 @@ class Api extends ApiKeyRequester {
         return response;
     }
 
-    async listAllWorkflows() {
+    async listAllWorkflows(params) {
         const options = {
             url: this.baseUrl + this.URLs.workflows,
+            query: params,
         };
         const response = await this._get(options);
         return response;
@@ -159,6 +165,26 @@ class Api extends ApiKeyRequester {
             body
         };
         const response = await this._patch(options);
+    async createWorkflowComment(id, body) {
+        const options = {
+            url: this.baseUrl + this.URLs.workflowComment(id),
+            headers: {
+                'content-type': 'application/json',
+            },
+            body,
+        };
+        const response = await this._post(options);
+        return response;
+    }
+
+    async retrieveWorkflowDocument(workflowID, documentKey) {
+        const options = {
+            url:
+                this.baseUrl +
+                this.URLs.workflowsByID(workflowID) +
+                `/document/${documentKey}/download`,
+        };
+        const response = await this._get(options);
         return response;
     }