+++ api-module-library/ironclad/CHANGELOG.md
@@ -1,3 +1,38 @@
+# v0.0.9 (Fri Oct 28 2022)
+
+#### 🐛 Bug Fix
+
+- Fr/create workflow comment [#63](https://github.com/friggframework/frigg/pull/63) ([@sheehantoufiq](https://github.com/sheehantoufiq))
+- merge changes ([@sheehantoufiq](https://github.com/sheehantoufiq))
+- Merge branch 'main' into fr/create-workflow-comment ([@sheehantoufiq](https://github.com/sheehantoufiq))
+- test bug fix ([@sheehantoufiq](https://github.com/sheehantoufiq))
+- testing bug ([@sheehantoufiq](https://github.com/sheehantoufiq))
+- update workflow ([@sheehantoufiq](https://github.com/sheehantoufiq))
+- create workflow comment ([@sheehantoufiq](https://github.com/sheehantoufiq))
+
+#### Authors: 1
+
+- Sheehan Toufiq Khan ([@sheehantoufiq](https://github.com/sheehantoufiq))
+
+---
+
+# v0.0.8 (Fri Oct 28 2022)
+
+#### 🐛 Bug Fix
+
+- Ironclad - Retrieve Workflow Document [#58](https://github.com/friggframework/frigg/pull/58) ([@JonathanEdMoore](https://github.com/JonathanEdMoore))
+- Fix: Ironclad List All Workflows [#59](https://github.com/friggframework/frigg/pull/59) ([@JonathanEdMoore](https://github.com/JonathanEdMoore))
+- Passing Tests ([@JonathanEdMoore](https://github.com/JonathanEdMoore))
+- Added params to listAllWorkflows ([@JonathanEdMoore](https://github.com/JonathanEdMoore))
+- tests Passing ([@JonathanEdMoore](https://github.com/JonathanEdMoore))
+- Added method to retrieve a workflow document ([@JonathanEdMoore](https://github.com/JonathanEdMoore))
+
+#### Authors: 1
+
+- Jonathan Moore ([@JonathanEdMoore](https://github.com/JonathanEdMoore))
+
+---
+
 # v0.0.7 (Wed Oct 19 2022)
 
 #### 🐛 Bug Fix